Progresses #38, added image scaling

+++ src/Interface/interface.py
@@ -36,7 +36,7 @@ class ZepazoParams(QMainWindow):
         height, width, ch = first_frame_qt.shape
         bytes_lines = ch * width
         qt_image = QtGui.QImage(first_frame_qt.data, width,height, bytes_lines, QtGui.QImage.Format_RGB888)
-        qt_image_scaled = qt_image.scaled(width/1.4, height/1.4, Qt.KeepAspectRatio)
+        qt_image_scaled = qt_image.scaled(int(width/1.4), int(height/1.4), Qt.KeepAspectRatio)
         self.centralPanel.setPixmap(QPixmap.fromImage(qt_image_scaled))
         self.centralPanel.setScaledContents(True)
